move reports app into data
